Change ho the AAD is set
